% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{decode_olc}
\alias{decode_olc}
\title{Decode Open Location Codes into Latitude and Longitude Pairs}
\usage{
decode_olc(olcs)
}
\arguments{
\item{olcs}{a vector of Open Location Codes, generated through \code{encode_olc} or
an equivalent tool.}
}
\description{
\code{decode_olc} takes Open Location Codes and, if they're
valid (see \code{\link{validate_full}}) returns the minium, centred and maximum
latitude and longitude for those coordinates.
}
\examples{
decode_olc("7FG49Q00+")

}
\seealso{
\code{\link{encode_olc}} for the opposite operation, and \code{\link{shorten_olc}} to convert
"full" Open Location Codes to "short" Open Location Codes.
}

