% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{valid_olc}
\alias{olc_validate}
\alias{valid_full}
\alias{valid_olc}
\alias{valid_short}
\title{Check the Validity of Open Location Codes}
\usage{
valid_olc(codes)

valid_short(codes)

valid_full(codes)
}
\arguments{
\item{codes}{a character vector containing Open Location Codes.}
}
\value{
a vector of TRUE and FALSE values, where TRUE corresponds to a
valid code and FALSE an invalid.
}
\description{
These functions allow a useR to check whether OLCs they've
been provided are valid or not. \code{valid_short} identifies whether
a vector of OLCs are valid "short" codes; \code{valid_long} identifies
whether OLCs are valid "long" codes, and \code{valid_full} identifies
whether OLCs are valid, full stop.
}

