% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olctools.R
\docType{package}
\name{olctools}
\alias{olctools}
\alias{olctools-package}
\title{Tools for handling Open Location Codes}
\description{
\href{http://openlocationcode.com/}{Open Location Codes} are a Google-created standard for identifying geographic locations.
olctools provides utilities for validating, encoding and decoding entries that follow this standard.
}

